requirements.txt
langchain==0.0.327
openai
faiss-cpu
streamlit
streamlit-chat
tiktoken
